requirements.txt
ConfigParser
schedule
paho-mqtt